fix: Solve form data not updating
Extract form submit logic from loanFormReducer to handleSubmit in Home.tsx. This is required because loanFormReducer should only handle actions related to manipulating the form data. Submitting data to an api is out of scope for the reducer. Also, add a timeout before getLoans request to updates are present in the provider.